Switch to a custom extraction script
More solid and easy to use extraction scripts make developers and
maintainers happy. This new script was heavily influenced by the CM
script in use by many projects, available inside device trees.

The main improvement over the old scripts is that now, only this
repository has to be cloned to be worked on. The device tree repository
can be cloned and a lookup for the proprietary-blobs.txt file is still
automatically done there, allowing the script to be used in a similar
manner as before, but now, you can grab it as a stand-alone as well,
meaning the script will also look for proprietary-blobs.txt files you
drop at the root of this repository. Rejoice!