Sort output of the extraction script
When something changes later down the line, we just get questions of
what exactly changed. Sometimes lines just magically move around in
the makefiles making it hard to understand that without following
the individual lines through. Starting with this revision of the
extraction script, we get sorted output to the makefiles.