Make the repository compatible with AOSP
The changes to the extraction script modify the .mk files generated
to follow the structure that the AOSP repo expects. This is great if
someone wants to grab a repo that has relevant blobs for an AOSP
build.

Our own device trees need small updates when switching over to this,
as can be seen in Change-ID I258f22cbd2f398f235c0888492d4283397ac2ea3
in the mako device tree, for example.